Prevent heap allocation that is never cleared
delete[] is never called for Decryption table, and is allocated to the heap, so will leak. Does not seem to be used outside of BlackOps4.cpp, removing the pointer, and instead allocating on the stack without the new keywork stops this memory leak and doesn't seem to break any of the functionality of it.